feat: add helm chart
This change adds a helm chart for `spotinst-metrics-exporter` and a
workflow which will automatically publish the chart to GitHub Pages
whenever there is a change in the `main` branch.

GitHub Pages was already enabled and the `gh-pages` branch exists.

This is identically to how we do it for `pod-image-swap-webhook` except
that the chart is a bit simpler since it does not require any of the
webhook-specific resources.

Have a look at the `README.md` to see how it's used.