feat: add support for propagating custom resource labels
The new CLI flag `--resource-labels` allows configuring Kubernetes
resource labels that should be propagated into Prometheus metric labels.

The flag can be specified multiple times and expects a comma-separated
list of Kubernetes resource label names with an optional mapping to a
Prometheus label.

Examples:

- `team` propagates the value of the Kubernetes label `team` into a
  Prometheus label of the same name.
- `app.kubernetes.io/name=app` propagates the value of the Kubernetes
  label `app.kubernetes.io/name` into the Prometheus label `app`.

If a resource label is configured but not available on a resource, the
Prometheus label's value will default to an empty string.

The following metrics benefit from this change:

- `spotinst_ocean_aws_namespace_cost`
- `spotinst_ocean_aws_workload_cost`

All other metrics (e.g. resource suggestions) do not benefit from this
change because there are no resource labels returned from the spotinst
API. Support could be added in the future, at the expense of additional
API calls to enrich the currently available data.

A helm chart change to support this feature will be provided in a
separate PR.